refactor(components): migrate HandCard, FieldCard, FieldSpellTrap, HoverPreview, CardActivationOverlay to new card views
